driver_actors functions moved to mln
